[feat] Added navigation bar for different pages.
